# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/cms/v1/room/availibility-room": {
            "get": {
                "description": "Room Get Availibility Room",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Room"
                ],
                "operationId": "RoomGetAvailibilityRoom",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer %",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "startDate",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "floorNumber",
                        "name": "floorNumber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "roomNumber",
                        "name": "roomNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "roomTypeName",
                        "name": "roomTypeName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "startFloorPrice",
                        "name": "startPrice",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "endfloorPrice",
                        "name": "endPrice",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CommonBaseResponse"
                        }
                    }
                }
            }
        },
        "/api/cms/v1/room/create": {
            "post": {
                "description": "Room Create",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Room"
                ],
                "operationId": "RoomCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer %",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "floorId",
                        "name": "floorId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "number",
                        "name": "number",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CommonBaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.AlertResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "inner_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.CommonBaseResponse": {
            "type": "object",
            "properties": {
                "alert": {
                    "$ref": "#/definitions/responses.AlertResponse"
                },
                "data": {}
            }
        }
    }
}